refactor(utils): move out 'cleanup' utils to a separate module
It will allow to be more flexible in importing other common utils
to avoid circular imports.
