+++ unit_tests/test_clean_cloud_resources_func.py
@@ -12,13 +12,19 @@
 # Copyright (c) 2020 ScyllaDB
 
 import unittest
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, Mock, patch
 
 from sdcm.sct_config import SCTConfiguration
-from sdcm.utils.common import \
-    clean_cloud_resources, \
-    clean_instances_aws, clean_elastic_ips_aws, clean_clusters_gke, clean_instances_gce, clean_resources_docker
 from sdcm.utils.context_managers import environment
+from sdcm.utils import resources_cleanup
+from sdcm.utils.resources_cleanup import (
+    clean_cloud_resources,
+    clean_clusters_gke,
+    clean_elastic_ips_aws,
+    clean_instances_aws,
+    clean_instances_gce,
+    clean_resources_docker,
+)
 
 
 SCT_RUNNER_AWS = {
@@ -33,21 +39,23 @@ class CleanInstanceAwsTest(unittest.TestCase):
         self.assertRaisesRegex(AssertionError, "not provided", clean_instances_aws, {})
 
     def test_empty_list(self, ec2_client):  # pylint: disable=no-self-use
-        with patch("sdcm.utils.common.list_instances_aws", return_value={}) as list_instances_aws:
-            clean_instances_aws({"TestId": 1111, })
+        with patch.object(resources_cleanup, "list_instances_aws", Mock(return_value={})) as list_instances_aws:
+            resources_cleanup.clean_instances_aws({"TestId": 1111, })
         list_instances_aws.assert_called_with(tags_dict={"TestId": 1111, }, group_as_region=True)
         ec2_client().terminate_instances.assert_not_called()
 
     def test_sct_runner(self, ec2_client):  # pylint: disable=no-self-use
-        with patch("sdcm.utils.common.list_instances_aws",
-                   return_value={"eu-north-1": [SCT_RUNNER_AWS, ]}) as list_instances_aws:
+        with patch.object(
+                resources_cleanup, "list_instances_aws",
+                Mock(return_value={"eu-north-1": [SCT_RUNNER_AWS]})) as list_instances_aws:
             clean_instances_aws({"TestId": 1111, })
         list_instances_aws.assert_called_with(tags_dict={"TestId": 1111, }, group_as_region=True)
         ec2_client().terminate_instances.assert_not_called()
 
     def test_terminate(self, ec2_client):  # pylint: disable=no-self-use
-        with patch("sdcm.utils.common.list_instances_aws",
-                   return_value={"eu-north-1": [{"InstanceId": "i-1111", }, ]}) as list_instances_aws:
+        with patch.object(
+                resources_cleanup, "list_instances_aws",
+                Mock(return_value={"eu-north-1": [{"InstanceId": "i-1111"}]})) as list_instances_aws:
             clean_instances_aws({"TestId": 1111, })
         list_instances_aws.assert_called_with(tags_dict={"TestId": 1111, }, group_as_region=True)
         ec2_client().terminate_instances.assert_called_once_with(InstanceIds=["i-1111"])
@@ -59,19 +67,23 @@ class CleanElasticIpsAws(unittest.TestCase):
         self.assertRaisesRegex(AssertionError, "not provided", clean_elastic_ips_aws, {})
 
     def test_empty_list(self, ec2_client):  # pylint: disable=no-self-use
-        with patch("sdcm.utils.common.list_elastic_ips_aws", return_value={}) as list_elastic_ips_aws:
+        with patch.object(
+                resources_cleanup, "list_elastic_ips_aws",
+                Mock(return_value={})) as list_elastic_ips_aws:
             clean_elastic_ips_aws({"TestId": 1111, })
         list_elastic_ips_aws.assert_called_with(tags_dict={"TestId": 1111, }, group_as_region=True)
         ec2_client().disassociate_address.assert_not_called()
         ec2_client().release_address.assert_not_called()
 
     def test_release(self, ec2_client):  # pylint: disable=no-self-use
-        with patch("sdcm.utils.common.list_elastic_ips_aws",
-                   return_value={"eu-north-1": [{
-                       "AssociationId": 2222,
-                       "AllocationId": 3333,
-                       "PublicIp": '127.0.0.1',
-                   }]}) as list_elastic_ips_aws:
+        return_value = {"eu-north-1": [{
+            "AssociationId": 2222,
+            "AllocationId": 3333,
+            "PublicIp": '127.0.0.1',
+        }]}
+        with patch.object(
+                resources_cleanup, "list_elastic_ips_aws",
+                Mock(return_value=return_value)) as list_elastic_ips_aws:
             clean_elastic_ips_aws({"TestId": 1111, })
         list_elastic_ips_aws.assert_called_with(tags_dict={"TestId": 1111, }, group_as_region=True)
         ec2_client().disassociate_address.assert_called_once_with(AssociationId=2222)
@@ -84,7 +96,9 @@ class CleanClustersGkeTest(unittest.TestCase):
 
     def test_destroy(self):  # pylint: disable=no-self-use
         cluster = MagicMock()
-        with patch("sdcm.utils.common.list_clusters_gke", return_value=[cluster, ]) as list_clusters_gke:
+        with patch.object(
+                resources_cleanup, "list_clusters_gke",
+                Mock(return_value=[cluster])) as list_clusters_gke:
             clean_clusters_gke({"TestId": 1111, })
         list_clusters_gke.assert_called_with(tags_dict={"TestId": 1111, })
         cluster.destroy.assert_called_once_with()
@@ -96,11 +110,13 @@ class CleanInstancesGceTest(unittest.TestCase):
 
     def test_destroy(self):  # pylint: disable=no-self-use
         instance = MagicMock()
-        with patch("sdcm.utils.common.list_instances_gce",
-                   return_value=[instance, ]) as list_instances_gce, \
-                patch('sdcm.utils.common.get_gce_compute_instances_client',
-                      return_value=(instance, dict(project_id='test'))):
-            clean_instances_gce({"TestId": 1111, })
+        with patch.object(
+                resources_cleanup, "list_instances_gce",
+                Mock(return_value=[instance])) as list_instances_gce:
+            with patch.object(
+                    resources_cleanup, 'get_gce_compute_instances_client',
+                    Mock(return_value=(instance, dict(project_id='test')))):
+                clean_instances_gce({"TestId": 1111, })
         list_instances_gce.assert_called_with(tags_dict={"TestId": 1111, })
         instance.delete.assert_called_once_with(instance=instance.name, project='test', zone=unittest.mock.ANY)
 
@@ -113,8 +129,9 @@ class CleanResourcesDockerTest(unittest.TestCase):
     def test_destroy():
         image = MagicMock()
         container = MagicMock()
-        with patch("sdcm.utils.common.list_resources_docker",
-                   return_value={"images": [image, ], "containers": [container, ], }) as list_resources_docker:
+        with patch.object(
+                resources_cleanup, "list_resources_docker",
+                Mock(return_value={"images": [image], "containers": [container]})) as list_resources_docker:
             clean_resources_docker({"TestId": 1111, })
         list_resources_docker.assert_called_with(
             tags_dict={"TestId": 1111, },
@@ -128,16 +145,16 @@ class CleanResourcesDockerTest(unittest.TestCase):
 class CleanCloudResourcesTest(unittest.TestCase):
     integration = False  # set it to True if you want to run test with actual cloud operations.
     functions_to_patch = (
-        "sdcm.utils.common.clean_instances_aws",
-        "sdcm.utils.common.clean_elastic_ips_aws",
-        "sdcm.utils.common.clean_clusters_gke",
-        "sdcm.utils.common.clean_orphaned_gke_disks",
-        "sdcm.utils.common.clean_clusters_eks",
-        "sdcm.utils.common.clean_instances_gce",
-        "sdcm.utils.common.clean_instances_azure",
-        "sdcm.utils.common.clean_resources_docker",
-        "sdcm.utils.common.clean_test_security_groups",
-        "sdcm.utils.common.clean_load_balancers_aws",
+        "sdcm.utils.resources_cleanup.clean_instances_aws",
+        "sdcm.utils.resources_cleanup.clean_elastic_ips_aws",
+        "sdcm.utils.resources_cleanup.clean_clusters_gke",
+        "sdcm.utils.resources_cleanup.clean_orphaned_gke_disks",
+        "sdcm.utils.resources_cleanup.clean_clusters_eks",
+        "sdcm.utils.resources_cleanup.clean_instances_gce",
+        "sdcm.utils.resources_cleanup.clean_instances_azure",
+        "sdcm.utils.resources_cleanup.clean_resources_docker",
+        "sdcm.utils.resources_cleanup.clean_test_security_groups",
+        "sdcm.utils.resources_cleanup.clean_load_balancers_aws",
     )
 
     @classmethod
